requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.23.0
pyyaml>=6.0